chore: move app into src folder for better organization
