+++ public/sitemap.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://www.christinewessa.com</loc><changefreq>daily</changefreq><priority>0.8</priority></url>
 <url><loc>https://www.christinewessa.com/contact-christine</loc><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.christinewessa.com</loc><changefreq>daily</changefreq><priority>0.8</priority></url>
 <url><loc>https://www.christinewessa.com/blog</loc><changefreq>daily</changefreq><priority>0.7</priority></url>
 <url><loc>https://www.christinewessa.com/blog/the-lord-doesn&apos;t-need-you</loc><changefreq>daily</changefreq><priority>0.7</priority></url>
 <url><loc>https://www.christinewessa.com/blog/let-yourself-be-seen</loc><changefreq>daily</changefreq><priority>0.7</priority></url>
