+++ src/app/(app)/home/HomeHero.tsx
