+++ src/app/components/ImageStamp.tsx
