+++ src/app/components/SocialIconsList.tsx
