+++ src/app/data/socialIconsData.tsx
