+++ src/app/types/blogTypes.ts
