+++ tsconfig.json
@@ -17,8 +17,9 @@
         "name": "next"
       }
     ],
+    "baseUrl": "src/app",
     "paths": {
-      "@/*": ["./*"]
+      "@/components/*": ["components/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
